Unit test for attachment and assignment
